+++ saleor/graphql/page/mutations/pages.py
@@ -27,10 +27,7 @@ if TYPE_CHECKING:
 class PageInput(graphene.InputObjectType):
     slug = graphene.String(description="Page internal name.")
     title = graphene.String(description="Page title.")
-    content = graphene.String(
-        description=("Page content. May consist of ordinary text, HTML and images.")
-    )
-    content_json = graphene.JSONString(description="Page content in JSON format.")
+    content = graphene.JSONString(description="Page content in JSON format.")
     attributes = graphene.List(
         graphene.NonNull(AttributeValueInput), description="List of attributes."
     )
@@ -111,6 +108,11 @@ class PageCreate(ModelMutation):
         if attributes:
             AttributeAssignmentMixin.save(instance, attributes)
 
+    @classmethod
+    def save(cls, info, instance, cleaned_input):
+        super().save(info, instance, cleaned_input)
+        info.context.plugins.page_created(instance)
+
 
 class PageUpdate(PageCreate):
     class Arguments:
@@ -126,6 +128,11 @@ class PageUpdate(PageCreate):
         error_type_class = PageError
         error_type_field = "page_errors"
 
+    @classmethod
+    def save(cls, info, instance, cleaned_input):
+        super(PageCreate, cls).save(info, instance, cleaned_input)
+        info.context.plugins.page_updated(instance)
+
 
 class PageDelete(ModelDeleteMutation):
     class Arguments:
@@ -137,6 +144,13 @@ class PageDelete(ModelDeleteMutation):
         permissions = (PagePermissions.MANAGE_PAGES,)
         error_type_class = PageError
         error_type_field = "page_errors"
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        page = cls.get_instance(info, **data)
+        response = super().perform_mutation(_root, info, **data)
+        info.context.plugins.page_deleted(page)
+        return response
 
 
 class PageTypeCreateInput(graphene.InputObjectType):