+++ saleor/graphql/product/mutations/products.py
@@ -12,6 +12,7 @@ from ....attribute import AttributeInputType, AttributeType
 from ....attribute import models as attribute_models
 from ....core.exceptions import PermissionDenied
 from ....core.permissions import ProductPermissions, ProductTypePermissions
+from ....core.utils.editorjs import clean_editor_js
 from ....order import OrderStatus
 from ....order import models as order_models
 from ....product import models
@@ -60,8 +61,7 @@ from ..utils import (
 
 
 class CategoryInput(graphene.InputObjectType):
-    description = graphene.String(description="Category description (HTML/text).")
-    description_json = graphene.JSONString(description="Category description (JSON).")
+    description = graphene.JSONString(description="Category description (JSON).")
     name = graphene.String(description="Category name.")
     slug = graphene.String(description="Category slug.")
     seo = SeoInput(description="Search engine optimization fields.")
@@ -171,10 +171,7 @@ class CollectionInput(graphene.InputObjectType):
     )
     name = graphene.String(description="Name of the collection.")
     slug = graphene.String(description="Slug of the collection.")
-    description = graphene.String(
-        description="Description of the collection (HTML/text)."
-    )
-    description_json = graphene.JSONString(
+    description = graphene.JSONString(
         description="Description of the collection (JSON)."
     )
     background_image = Upload(description="Background image file.")
@@ -484,8 +481,7 @@ class ProductInput(graphene.InputObjectType):
         description="List of IDs of collections that the product belongs to.",
         name="collections",
     )
-    description = graphene.String(description="Product description (HTML/text).")
-    description_json = graphene.JSONString(description="Product description (JSON).")
+    description = graphene.JSONString(description="Product description (JSON).")
     name = graphene.String(description="Product name.")
     slug = graphene.String(description="Product slug.")
     tax_code = graphene.String(description="Tax rate for enabled tax gateway.")
@@ -539,6 +535,11 @@ class ProductCreate(ModelMutation):
     def clean_input(cls, info, instance, data):
         cleaned_input = super().clean_input(info, instance, data)
 
+        description = cleaned_input.get("description")
+        cleaned_input["description_plaintext"] = (
+            clean_editor_js(description, to_string=True) if description else ""
+        )
+
         weight = cleaned_input.get("weight")
         if weight and weight.value < 0:
             raise ValidationError(
@@ -567,11 +568,6 @@ class ProductCreate(ModelMutation):
         except ValidationError as error:
             error.code = ProductErrorCode.REQUIRED.value
             raise ValidationError({"slug": error})
-
-        # FIXME  tax_rate logic should be dropped after we remove tax_rate from input
-        tax_rate = cleaned_input.pop("tax_rate", "")
-        if tax_rate:
-            info.context.plugins.assign_tax_code_to_object_meta(instance, tax_rate)
 
         if "tax_code" in cleaned_input:
             info.context.plugins.assign_tax_code_to_object_meta(
@@ -611,7 +607,6 @@ class ProductCreate(ModelMutation):
     @transaction.atomic
     def save(cls, info, instance, cleaned_input):
         instance.save()
-
         attributes = cleaned_input.get("attributes")
         if attributes:
             AttributeAssignmentMixin.save(instance, attributes)
@@ -623,10 +618,13 @@ class ProductCreate(ModelMutation):
             instance.collections.set(collections)
 
     @classmethod
+    def post_save_action(cls, info, instance, cleaned_input):
+        info.context.plugins.product_created(instance)
+
+    @classmethod
     def perform_mutation(cls, _root, info, **data):
         response = super().perform_mutation(_root, info, **data)
         product = getattr(response, cls._meta.return_field_name)
-        info.context.plugins.product_created(product)
 
         # Wrap product instance with ChannelContext in response
         setattr(
@@ -658,6 +656,9 @@ class ProductUpdate(ProductCreate):
         attributes = cleaned_input.get("attributes")
         if attributes:
             AttributeAssignmentMixin.save(instance, attributes)
+
+    @classmethod
+    def post_save_action(cls, info, instance, cleaned_input):
         info.context.plugins.product_updated(instance)
 
 
@@ -680,19 +681,19 @@ class ProductDelete(ModelDeleteMutation):
     @classmethod
     def perform_mutation(cls, _root, info, **data):
         node_id = data.get("id")
-        instance = cls.get_node_or_error(info, node_id, only_type=Product)
 
+        instance = cls.get_node_or_error(info, node_id, only_type=Product)
+        variants_id = list(instance.variants.all().values_list("id", flat=True))
         # get draft order lines for variant
         line_pks = list(
             order_models.OrderLine.objects.filter(
-                variant__in=instance.variants.all(), order__status=OrderStatus.DRAFT
+                variant_id__in=variants_id, order__status=OrderStatus.DRAFT
             ).values_list("pk", flat=True)
         )
-
         response = super().perform_mutation(_root, info, **data)
-
         # delete order lines for deleted variant
         order_models.OrderLine.objects.filter(pk__in=line_pks).delete()
+        info.context.plugins.product_deleted(instance, variants_id)
 
         return response
 
@@ -1063,14 +1064,6 @@ class ProductTypeCreate(ModelMutation):
         except ValidationError as error:
             error.code = ProductErrorCode.REQUIRED.value
             raise ValidationError({"slug": error})
-
-        # FIXME  tax_rate logic should be dropped after we remove tax_rate from input
-        tax_rate = cleaned_input.pop("tax_rate", "")
-        if tax_rate:
-            instance.store_value_in_metadata(
-                {"vatlayer.code": tax_rate, "description": tax_rate}
-            )
-            info.context.plugins.assign_tax_code_to_object_meta(instance, tax_rate)
 
         tax_code = cleaned_input.pop("tax_code", "")
         if tax_code: