+++ saleor/settings.py
@@ -247,7 +247,6 @@ INSTALLED_APPS = [
     "saleor.invoice",
     "saleor.seo",
     "saleor.shipping",
-    "saleor.search",
     "saleor.site",
     "saleor.page",
     "saleor.payment",
@@ -466,10 +465,10 @@ PLACEHOLDER_IMAGES = {
 
 DEFAULT_PLACEHOLDER = "images/placeholder255x255.png"
 
-SEARCH_BACKEND = "saleor.search.backends.postgresql"
 
 AUTHENTICATION_BACKENDS = [
     "saleor.core.auth_backend.JSONWebTokenBackend",
+    "saleor.core.auth_backend.PluginBackend",
 ]
 
 # CELERY SETTINGS